requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
opencv-python>=4.5.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.60.0
Pillow>=8.0.0
albumentations>=1.1.0
numpy>=1.21.0
